Refactor part of SchedulePage component to make date-specific filters independent of each other.
